requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0
pandas>=2.0.0
sqlalchemy>=2.0.0
h3>=4.0.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
pydantic>=2.0.0
thefuzz>=0.22.0